requirements-freeze.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
archspec==0.2.5
argcomplete==3.6.3
attrs==25.1.0
audioop-lts==0.2.2
audioread==3.1.0
av==16.0.1
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
Beaker==1.12.1
beautifulsoup4==4.14.3
bitsandbytes==0.47.0
blivet==3.12.1
blivet-gui==2.6.0
boltons==25.0.0
boto3==1.40.32
botocore==1.40.32
Brlapi==0.8.6
build==1.3.0
cachetools==5.5.2
camelot-py==1.0.9
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.0.21
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-351-build/cockpit-351/tmp/wheel/cockpit-351-py3-none-any.whl#sha256=59d2dcd42c86f8d88df2a0b4afd6bee5c58753efde42f78d6ed1686b14aa61df
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
conda==24.11.3
conda-libmamba-solver==24.9.0
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
contourpy==1.3.3
crcmod==1.7
crypt_r==3.13.1
cryptography==44.0.0
ctranslate2==4.6.2
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dataclasses-json==0.6.7
datasets==4.1.0
dbus-python==1.3.2
dbus_next==0.2.3
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
dirtyjson==1.0.8
distro==1.9.0
dnf==4.24.0
dnspython==2.7.0
docopt==0.6.2
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
evdev==1.9.1
eyeD3==0.9.9
faiss-cpu==1.12.0
fastapi==0.115.14
fastapi-slim==0.115.14
faster-whisper==1.2.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.60.1
fros==1.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
gbinder-python==1.3.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hidapi==0.14.0.post3
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.5
humanfriendly==10.0
humanize==4.12.0
hyperframe==6.1.0
idna==3.10
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
iso639==0.1.4
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jq==1.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
kubernetes==33.1.0
labelImg==1.8.6
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.29
langchain-core==0.3.76
langchain-ollama==0.3.8
langchain-text-splitters==0.3.11
langsmith==0.4.28
langtable==0.0.69
lazy_loader==0.4
libcomps==0.1.22
libdnf==0.75.0
libmambapy==1.5.12
librosa==0.11.0
libvirt-python==11.0.0
liquidctl==1.15.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.1
llama-index-cli==0.5.1
llama-index-core==0.14.2
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.2.0
llama-parse==0.6.54
llvmlite==0.46.0
looseversion==1.3.0
louis==3.33.0
lutris==0.5.19
lxml==5.3.2
Mako==1.2.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.31
menuinst==2.0.0
mmh3==5.2.0
moddb==0.12.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nftables==0.1
nltk==3.9.1
numba==0.63.1
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
olefile==0.47
ollama==0.5.4
onnxruntime==1.22.1
openai==1.107.3
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
Paste==3.10.1
pdf2image==1.17.0
pdfminer.six==20251107
peft==0.17.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.4.0
pluggy==1.6.0
ply==3.11
polars==1.35.2
polars-runtime-32==1.35.2
pooch==1.8.2
portalocker==3.2.0
posthog==5.4.0
productmd==1.49
proglog==0.1.12
progressbar2==4.5.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pwquality==1.4.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.13
pyAudioAnalysis==0.3.14
pybase64==1.4.2
pycairo==1.25.1
pyclip==0.7.0
pycosat==0.6.6
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyenchant==3.2.2
Pygments==2.19.2
PyGObject==3.50.2
pyinotify==0.9.6
pykickstart==3.62
pylance==0.36.0
pyOpenSSL==25.0.0
pyparsing==3.2.5
pyparted==3.13.0
pypdf==5.9.0
pypdfium2==5.0.0
PyPika==0.48.9
pypresence==4.3.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
PySocks==1.7.1
pysqlite3-binary==0.5.4.post2
pytesseract==0.3.13
pytest==8.4.2
python-augeas==1.2.0
python-dateutil==2.8.2
python-dotenv==1.1.1
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.9.0
pytz==2025.2
pyudev==0.24.3
pyusb==1.3.1
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.17.13-build/kernel-6.17.13/linux-6.17.13-200.fc42.x86_64/tools/net/ynl
qdrant-client==1.15.1
RapidFuzz==3.11.0
referencing==0.36.2
regex==2025.11.3
reportlab==4.4.3
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.25.0
rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.12
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sdkmanager==0.6.10
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sentence-transformers==5.1.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setproctitle==1.3.6
setuptools==80.9.0
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
smbus==1.1
sniffio==1.3.1
sos==4.10.0
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.7
soxr==1.0.0
SQLAlchemy==2.0.43
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.47.3
striprtf==0.0.26
sympy==1.13.1
systemd-python==235
tabulate==0.9.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.5.1+cu121
tqdm==4.67.1
transformers==4.56.1
triton==3.5.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.229
ultralytics-thop==2.0.18
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
whisper==1.1.10
wrapt==1.17.3
xkbregistry==0.3
xxhash==3.5.0
yarl==1.20.1
yt-dlp==2025.12.8
zipp==3.23.0
zstandard==0.25.0
